refactor: move AggregationResult to core package
