+++ search/search-client-elasticsearch/src/main/java/io/camunda/search/es/transformers/aggregator/SearchAggregationResultTransformer.java
@@ -13,8 +13,8 @@ import co.elastic.clients.elasticsearch._types.aggregations.MultiBucketAggregate
 import co.elastic.clients.elasticsearch._types.aggregations.MultiBucketBase;
 import co.elastic.clients.elasticsearch._types.aggregations.SingleBucketAggregateBase;
 import co.elastic.clients.elasticsearch._types.aggregations.StringTermsBucket;
-import io.camunda.search.clients.aggregator.AggregationResult;
-import io.camunda.search.clients.aggregator.AggregationResult.Builder;
+import io.camunda.search.clients.core.AggregationResult;
+import io.camunda.search.clients.core.AggregationResult.Builder;
 import io.camunda.search.clients.transformers.SearchTransfomer;
 import java.util.HashMap;
 import java.util.List;
