+++ search/search-client-query-transformer/src/main/java/io/camunda/search/clients/transformers/aggregation/result/AggregationResultTransformer.java
@@ -8,7 +8,7 @@
 package io.camunda.search.clients.transformers.aggregation.result;
 
 import io.camunda.search.aggregation.result.AggregationResultBase;
-import io.camunda.search.clients.aggregator.AggregationResult;
+import io.camunda.search.clients.core.AggregationResult;
 import io.camunda.search.clients.transformers.ServiceTransformer;
 import java.util.Map;
